+++ src/main/java/wooteco/subway/station/dao/StationDao.java
@@ -6,7 +6,6 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.line.domain.Line;
 import wooteco.subway.station.application.StationException;
 import wooteco.subway.station.domain.Station;
 
@@ -23,14 +22,6 @@ public class StationDao {
             new Station(
                     rs.getLong("id"),
                     rs.getString("name")
-            );
-
-    private RowMapper<Line> lineInfoRowMapper = (rs, rowNum) ->
-            new Line(
-                    rs.getLong("id"),
-                    rs.getString("name"),
-                    rs.getString("color"),
-                    rs.getInt("extra_fare")
             );
 
     public StationDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
@@ -64,7 +55,7 @@ public class StationDao {
     public Optional<Station> findById(Long id) {
         try {
             String sql = "select * from STATION where id = ?";
-            return Optional.of(jdbcTemplate.queryForObject(sql, rowMapper, id));
+            return Optional.ofNullable(jdbcTemplate.queryForObject(sql, rowMapper, id));
         } catch (Exception e) {
             return Optional.empty();
         }
@@ -73,12 +64,6 @@ public class StationDao {
     public boolean isExistName(String name) {
         String sql = "SELECT EXISTS (SELECT id FROM STATION WHERE name = ?)";
         return jdbcTemplate.queryForObject(sql, Boolean.class, name);
-    }
-
-    public List<Line> findLinesPassing(Station station) {
-        String sql = "SELECT DISTINCT L.id as line_id, L.name as line_name, L.color as line_color, L.extra_fare as extra_fare FROM LINE L " +
-                "left outer join SECTION S WHERE L.id = S.line_id AND (S.up_station_id =? OR S.down_station_id = ?)";
-        return jdbcTemplate.query(sql, lineInfoRowMapper, station.getId(), station.getId());
     }
 
     public void updateById(Long id, String name) {
